backup_20250305/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.8
comm==0.2.2
contourpy==1.2.1
cssselect==1.2.0
cycler==0.12.1
dateparser==1.2.1
debugpy==1.8.5
decorator==5.1.1
demjson3==3.0.6
distro==1.9.0
exceptiongroup==1.2.2
executing==2.0.1
facebook-scraper==0.2.59
fake-useragent==2.0.3
filelock==3.17.0
fonttools==4.53.1
frozendict==2.4.4
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.4
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multitasking==0.0.11
narwhals==1.29.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.0.1
oauthlib==3.2.2
openai==1.65.2
packaging==24.1
pandas==2.2.2
pandas_ta==0.3.14b0
parse==1.20.2
parso==0.8.4
peewee==3.17.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==6.0.0
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.47
protobuf==5.29.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyee==11.1.1
Pygments==2.18.0
pyparsing==3.1.4
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-html==0.10.0
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.23.1
safetensors==0.5.3
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
snscrape @ git+https://github.com/JustAnotherArchivist/snscrape.git@614d4c2029a62d348ca56598f87c425966aaec66
soupsieve==2.6
stack-data==0.6.3
streamlit==1.42.2
sympy==1.13.3
tenacity==9.0.0
tokenizers==0.21.0
toml==0.10.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
tweepy==4.15.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.3
update-checker==0.18.0
urllib3==1.26.20
w3lib==2.3.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
yfinance==0.2.43
zipp==3.20.0
